# Improve performance of Arrhenius rate evaluations