--- a/src/kinetics/InterfaceKinetics.cpp
+++ b/src/kinetics/InterfaceKinetics.cpp
@@ -548,6 +548,9 @@ bool InterfaceKinetics::addReaction(shared_ptr<Reaction> r_base, bool resize)
         // Add reaction rate to evaluator
         size_t index = m_interface_types[rate->type()];
         m_interface_rates[index]->add(nReactions() - 1, *rate);
+        if (resize) {
+            m_interface_rates[index]->resize(nReactions(), nTotalSpecies());
+        }
 
     } else if (r_base->reaction_type == BMINTERFACE_RXN) {
         throw NotImplementedError("InterfaceKinetics::addReaction");