@ischoegl 2022-03-17
This is an interesting concept. Presumably, vectorization could be done for each of the other `ReactionRate` specialization also? Fwiw, my recent tests on #1219 indicate that there should be gains whenever `updateFromStruct` can be avoided. Given that we have a lot of features already accumulated in 2.6.0b4, I'm wondering whether to include this now or for 3.0.

As an aside, I'm surprised that your speed tests put the CTI/XML ahead of YAML. I currently have YAML with an edge of around 4% on `main`. It's likely compiler/hardware dependent, though.