@speth 2022-08-26
> @speth … do you think that the speedup arises from vectorization via Eigen or from omitting calculations? If I understand correctly, only the former creates complications for the test you ran in the context of [#1211 (comment)](https://github.com/Cantera/cantera/issues/1211#issuecomment-1065991448)?

As I said in the initial PR description, the bulk of the benefit seems to come from skipping rate updates for rates that are constant.