@speth 2025-10-31
I'm closing this as it's now horribly out of date and I have some different (hopefully better) ideas on how to accelerate these calculations.